refact: refact of all current code
