+++ src/Domain/Quizz/Entity/Quizz.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace App\Entity;
+namespace App\Domain\Quizz\Entity;
 
 use Doctrine\ORM\Mapping as ORM;
-use App\Repository\QuizzRepository;
+use App\Domain\Quizz\Repository\QuizzRepository;
 use Doctrine\Common\Collections\Collection;
 use ApiPlatform\Core\Annotation\ApiResource;
 use Symfony\Component\HttpFoundation\File\File;
