+++ src/Domain/Quizz/Repository/CategoryRepository.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace App\Repository;
+namespace App\Domain\Quizz\Repository;
 
-use App\Entity\Category;
+use App\Domain\Quizz\Entity\Category;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
 use Doctrine\Persistence\ManagerRegistry;
