+++ src/Infrastructure/Social/Authenticator/AbstractSocialAuthenticator.php
